Creates test_core folder
